PR review feedback: simplify merge logic, fix period save, extract JS
- Remove _save_period placement sync (don't update decoupled placements)
- Use start_date/end_date properties instead of specific_* fields
- Simplify merge_group: remove dry_run, action list, org/extra_info merge
- Move merge logic from merge_assignments.py into services/assignments.py
- Regenerate migrations (fix conflicting 0006 numbering)
- Add (migration) tag to changelog
- Extract inline JS to static/js/placement_period_toggle.js with data attrs
